task 4

+++ task4/src/main/kotlin/io/github/eaxdev/model/Person.kt
@@ -2,6 +2,7 @@ package io.github.eaxdev.model
 
 import org.springframework.data.annotation.Id
 import org.springframework.data.elasticsearch.annotations.Document
+import java.math.BigDecimal
 
 @Document(indexName = "person")
 data class Person(
@@ -10,6 +11,6 @@ data class Person(
     val docid: String,
     val fio: String,
     val birthday: String,
-    val salary: Double,
+    val salary: BigDecimal,
     val gender: String
 )
